pyproject.toml
[project]
name = "haasappdev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.2",
    "python-dotenv>=1.1.1",
]
